update and reorganize requests
