+++ app/templates/days.html
@@ -1,0 +1,5 @@
+{% for day in days %}
+{ Day: {{day.temperature}},
+Date: {{day.date}},
+High Risk: {{day.high_risk}}}
+{% endfor %}
